requirements.txt
Django==4.2.11
djangorestframework==3.14.0
drf-yasg==1.21.7
celery==5.3.6
django-celery-results==2.5.1
gunicorn==22.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.7.0
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.1